Transfer input logic from QuestDummy to input System
The keys to accept and reject as well as to activate the QuestDummys, are passed into the input system.
In Entity there is now a Boolean that can be used for QuestDummy and eventual decision events.
For example, you could set it up to ask if the player wants to press a button or something similar.
Two strategies have now been added to the input system for this purpose. Accept with J for Yes and N for No.
In the strategies is queried whether there is an entity in the vicinity of the player that can be accepted.

Activating the Quest has also got its own strategy. This works in principle just as before, except that the LogMessage has been embellished.